[pillar] Revamp pillar file structure & allow arbitrary keys in datadog.yaml (#35)
Separates the root pillar object 'datadog' into two sub-objects:
- "config", which describes the contents of the datadog.yaml config file
that will be installed on the minions,
- "install_settings", which contains various customizable parameters
for the Agent install, such as the Agent version.

With this new structure, any arbitrary key can be put in "config",
and will be reflected in the minions' config files.

Agent 5 is using a .ini config file, and there's no ini renderer available easily, so we're using the old way with explicitly supported options (except a little cleaner: we're not matching regexes in a file we don't have control of). To comply with what was previously supported on Agent 5 on SaltStack, the api_key option is supported for now.