Fix Agent 6 & 7 install from beta repositories (#42)
Fixes the yum repository path for beta installs (Agent 6 betas are in beta/6, Agent 7 betas in beta/7).
Fixes the regex which checks if a beta version will be installed (there was a -[0-9]+ too much at the end of the regex). The previous regex would match 7.15.0~beta.1-1 and not 7.15.0~beta.1, which makes the install fail since we add the -1 again during the installation (giving 7.15.0~beta.1-1-1, which is not an existing version).